+++ packages/brick_graphql_generators/test/graphql_model_serdes_generator/test_annotation_subfields.dart
@@ -1,0 +1,71 @@
+import 'package:brick_graphql/graphql.dart';
+
+final output = r'''
+// GENERATED CODE DO NOT EDIT
+// This file should NOT be version controlled and should not be manually edited.
+part of '../brick.g.dart';
+
+Future<ToFromJson> _$ToFromJsonFromGraphql(Map<String, dynamic> data,
+    {required GraphqlProvider provider,
+    GraphqlFirstRepository? repository}) async {
+  return ToFromJson(assoc: ToFromJsonAssoc.fromJson(data['assoc'] as String));
+}
+
+Future<Map<String, dynamic>> _$ToFromJsonToGraphql(ToFromJson instance,
+    {required GraphqlProvider provider,
+    GraphqlFirstRepository? repository}) async {
+  return {'assoc': instance.assoc.toJson()};
+}
+
+/// Construct a [ToFromJson]
+class ToFromJsonAdapter extends GraphqlFirstAdapter<ToFromJson> {
+  ToFromJsonAdapter();
+
+  @override
+  final Map<String, RuntimeGraphqlDefinition> fieldsToGraphqlRuntimeDefinition =
+      {
+    'assoc': const RuntimeGraphqlDefinition(
+      association: false,
+      documentNodeName: 'assoc',
+      iterable: false,
+      subfields: <String>{'subfield1', 'subfield2'},
+      type: Map,
+    )
+  };
+
+  @override
+  Future<ToFromJson> fromGraphql(Map<String, dynamic> input,
+          {required provider,
+          covariant GraphqlFirstRepository? repository}) async =>
+      await _$ToFromJsonFromGraphql(input,
+          provider: provider, repository: repository);
+  @override
+  Future<Map<String, dynamic>> toGraphql(ToFromJson input,
+          {required provider,
+          covariant GraphqlFirstRepository? repository}) async =>
+      await _$ToFromJsonToGraphql(input,
+          provider: provider, repository: repository);
+}
+''';
+
+class ToFromJsonAssoc {
+  final int? integer;
+
+  ToFromJsonAssoc({
+    this.integer,
+  });
+
+  Map<String, int?> toJson() => {'integer': integer};
+
+  factory ToFromJsonAssoc.fromJson(String data) => ToFromJsonAssoc(integer: int.tryParse(data));
+}
+
+@GraphqlSerializable()
+class ToFromJson {
+  @Graphql(subfields: {'subfield1', 'subfield2'})
+  final ToFromJsonAssoc assoc;
+
+  ToFromJson({
+    required this.assoc,
+  });
+}